+++ reconfigurablePetri/undostack.h
@@ -14,7 +14,7 @@ class UndoStack:public QUndoStack
 public slots:
     void arcInserted(QGraphicsItem * item1, QGraphicsItem * item2,
                         const QPainterPath &arc_path,
-                    const QString &id, PTNscene * scene,int ArcWeight);
+                    const QString &id, PTNscene * scene,double ArcWeight);
     void itemMoved (QGraphicsItem * movedItem, const QPointF &oldPosition);
         void nodeRemoved (QGraphicsItem * item, PTNscene * scene);
         void arcRemoved (QGraphicsItem * item, PTNscene * scene);
